requirements.txt
pygsheets == 2.0.2
discord.py == 1.3.1
